Improve macOS packaging, hardening, and robustness
Add a macOS Debug workflow and VM testing docs, tighten macOS package checks, and land multiple safety/hardening fixes.

Highlights:
- Add .github/workflows/macos-debug.yml to build/stage OpenGL/Metal variants and publish diagnostics/artifacts.
- Add docs-dev/macos-vm-testing-workflow.md and helper PowerShell/bash guest scripts for compliant Apple-hardware VM/host testing; update AGENTS.md to reference the macOS workflow and inventory guidance.
- Strengthen manual release validation (.github/workflows/manual-release.yml): require PkgInfo, verify Mach-O arch with lipo, verify ad-hoc codesign signatures, validate game-module install_name and additional plist entries.
- Meson changes: set macOS game-module install_name to @loader_path/..., add platform-specific compile/link hardening flags for Darwin and propagate link hardening to game modules.
- CI/test updates: include tools/tests/filesystem_case_segments.py in py_compile and test runs (commit-validation and push-verification).
- Multiple code robustness and safety fixes across the codebase: safer string handling, bounded buffers, overflow checks, safer decl/material handling, improved logging buffer handling, and other defensive fixes to reduce crashes and unsafe behavior.

This commit groups build/config changes, CI/test inclusions, packaging validation, documentation, and many small source fixes to improve macOS support and general runtime safety.

+++ src/idlib/containers/Queue.h
@@ -2,6 +2,8 @@
 #ifndef __QUEUE_H__
 #define __QUEUE_H__
 
+#include <stddef.h>
+
 /*
 ===============================================================================
 
@@ -86,7 +88,7 @@ void idQueue<type, nodePtr>::Test(void) {
 	delete element;
 }
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 class idQueueTemplate {
 public:
 							idQueueTemplate( void );
@@ -101,12 +103,12 @@ private:
 
 #define QUEUE_NEXT_PTR( element )		(*((type**)(((byte*)element)+nextOffset)))
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 idQueueTemplate<type,nextOffset>::idQueueTemplate( void ) {
 	first = last = NULL;
 }
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 void idQueueTemplate<type,nextOffset>::Add( type *element ) {
 	QUEUE_NEXT_PTR(element) = NULL;
 	if ( last ) {
@@ -117,7 +119,7 @@ void idQueueTemplate<type,nextOffset>::Add( type *element ) {
 	last = element;
 }
 
-template< class type, int nextOffset >
+template< class type, size_t nextOffset >
 type *idQueueTemplate<type,nextOffset>::Get( void ) {
 	type *element;
