+++ src/idlib/rvMemSys.cpp
@@ -945,7 +945,7 @@ void *Mem_Alloc16( const int size, const char *fileName, const int lineNumber, b
 {
 	void *mem = Mem_AllocDebugMemory( size, fileName, lineNumber, true, tag );
 	// make sure the memory is 16 byte aligned
-	assert( ( ((int)mem) & 15) == 0 );
+	assert( ( ((uintptr_t)mem) & 15 ) == 0 );
 	return mem;
 }
 
@@ -960,7 +960,7 @@ void Mem_Free16( void *ptr, const char *fileName, const int lineNumber )
 		return;
 	}
 	// make sure the memory is 16 byte aligned
-	assert( ( ((int)ptr) & 15) == 0 );
+	assert( ( ((uintptr_t)ptr) & 15 ) == 0 );
 	Mem_FreeDebugMemory( ptr, fileName, lineNumber, true );
 }
